% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecode.R
\name{.}
\alias{.}
\alias{get_root}
\title{Simulate Live Coding}
\usage{
.(n = NULL, root = teach::get_root(), script = sprintf("\%s.R",
  sprintf("\%02d", n)))

get_root(path = getOption("teach_path"))
}
\arguments{
\item{n}{A script number within the root.}

\item{root}{A root folder.}

\item{script}{A script name.}

\item{path}{A path for the root.}
}
\description{
Use the \pkg{xfun} package to simulate live coding.

Use the \pkg{xfun} package to simulate live coding.
}
\examples{
get_root()
options(teach_path = '.bin/scripts/01')
get_root()
}
